refactor(layout): move header and footer in their own components
Close #359
