+++ components/main-footer/template.pug
@@ -1,0 +1,22 @@
+footer.footer.has-text-centered
+  div.container
+    div.columns
+      div.column.is-8-desktop.is-offset-2-desktop
+        nav.breadcrumb.has-dot-separator.is-centered.is-marginless(aria-label="breadcrumbs")
+          ul
+            li(v-for="footerLink in footerLinks" :class="{ 'is-active': isActive(footerLink.link) }")
+              nuxt-link(:to="localePath({ name: footerLink.link })" exact) {{$t('footer_menu.' + footerLink.key)}}
+        p
+          small(v-html="$t('components.site_layout.footer.license_html')")
+        p {{$t('components.site_layout.footer.version', {version: siteVersion})}}
+
+    div.columns
+      div.column.is-8-desktop.is-offset-2-desktop
+        div
+          a(href="https://vercel.com/?utm_source=foe-tools&utm_campaign=oss" target="_blank")
+            figure.image.is-100xauto.has-text-centered.is-centered(style="margin: auto")
+              figcaption Powered by
+              img(:src="$store.get('isDarkTheme') ? '/vercel-light.svg' : '/vercel-dark.svg'")
+
+  a.cookie-consent-setting(@click.prevent="$cookieConsent.show")
+    img(src="~assets/cookie.png" width="36" height="36")
