+++ components/main-header/components/dialogDayNight/DialogDayNight.vue
@@ -1,0 +1,3 @@
+<template src="./template.pug" lang="pug"></template>
+
+<script src="./script.js"></script>
