+++ components/main-header/components/dialogDayNight/script.js
