+++ components/main-header/components/dialogGlobalSettings/DialogGlobalSettings.vue
@@ -1,0 +1,3 @@
+<template src="./template.pug" lang="pug"></template>
+
+<script src="./script.js"></script>
