+++ components/main-header/components/dialogGlobalSettings/script.js
