+++ jest.config.js
@@ -24,8 +24,16 @@ module.exports = {
   collectCoverageFrom: [
     "**/*.{js,vue}",
     "!**/components/graph-canvas/**",
+    "!**/app/**",
     "!**/components/promotion-message-builder/**",
     "!**/components/import-promotion-message/**",
+    "!**/components/main-footer/**",
+    "!**/components/main-header/components/**",
+    "!**/components/sortable-item/**",
+    "!**/components/sortable-list/**",
+    "!**/modules/**",
+    "!**/static/**",
+    "!**/tmp/**",
     "!**/coverage/**",
     "!**/dist/**",
     "!**/layouts/*.vue",
